+++ server/src/routes/middleware/api/groups.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from 'express';
-import dbManager from '../../service/dbManager';
+import dbManager from '../../../service/dbManager';
 
 const groups = {
   all: async (req: Request, res: Response, next: NextFunction) => {
