+++ server/src/routes/middleware/api/index.ts
@@ -1,0 +1,21 @@
+import isLogin from './isLogin';
+import users from './users';
+import posts from './posts';
+import likes from './likes';
+import image from './image';
+import comments from './comments';
+import problems from './problems';
+import groups from './groups';
+import profile from './profile';
+
+export {
+  isLogin,
+  users,
+  posts,
+  likes,
+  image,
+  comments,
+  problems,
+  groups,
+  profile
+};
