+++ server/src/routes/middleware/api/posts.ts
@@ -1,8 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 
-import dbManager from '../../service/dbManager';
-import { pictureCheck } from '../../service/pictureCheck';
-import { PostAddData, PostUpdateData } from '../../types/interface';
+import dbManager from '../../../service/dbManager';
+import { pictureCheck } from '../../../service/pictureCheck';
+import { PostAddData, PostUpdateData } from '../../../types/interface';
 
 const posts = {
   get: async (
