+++ server/src/routes/middleware/oauth/index.ts
@@ -1,0 +1,4 @@
+import githubLogin from './login';
+import logout from './logout';
+
+export { githubLogin, logout };
