+++ server/src/routes/middleware/oauth/logout.ts
